Moved logic from UnitOfWork to IdentifierConverter
